+++ Changes.md
@@ -12,3 +12,4 @@ Version 3.3.0.1
   6. Added all 'showsection0' options, being 'hide', 'show' and 'show only section 0'.
   7. Added 'Edit course settings' to the course administration menu.
   8. Disable 'sectioncontentjustification' setting.
+  9. Remove 'Your progress' and add down arrow.